chore(jarvis): audit ruff safe fixes - 6 items (-4 vs strict 66)
Audit dette 2026-05-27 nuit, priorite 2 (fixes safe sans risque) :
- UP017  : datetime.timezone.utc -> datetime.UTC (security_whitelists)
- SIM114 : if/elif meme action fusionnes via 'or' (bypass/filesystem + proxmox)
- RUF046 : int(round(x)) redondant -> round(x) (settings/routes)
- C416   : dict comprehension triviale -> dict() (system/routes)
- SIM115 : noqa + commentaire (voice/deepfilter open /dev/null
           ferme explicitement en finally L. 82, mutation globale sys.stderr
           incompatible avec context manager)

ruff strict 66 -> 62 errors. Ruff config projet : All checks passed.
1331 pytest pass (+37 vs baseline 1294).

+++ scripts/settings/routes.py
@@ -294,7 +294,7 @@ def api_set_dsp_params():
         elif isinstance(orig, (int, float)) and isinstance(v, (int, float)):
             lo, hi = _dsp_bounds.get(k, (-9999, 9999))
             clamped = max(lo, min(hi, v))
-            sanitized[k] = int(round(clamped)) if isinstance(orig, int) else round(float(clamped), 6)
+            sanitized[k] = round(clamped) if isinstance(orig, int) else round(float(clamped), 6)
         elif isinstance(orig, str):
             sanitized[k] = str(v)[:128]
     dsp_params.update(sanitized)
